move helper function to networf_def file
